Got bulk adding season requests to be semi working

+++ api/src/main/java/com/clubhub/entity/Records/UserSeason.java
@@ -28,13 +28,13 @@ public class UserSeason {
     private User player;
 
     @Column
-    private int season;
+    private Long season;
 
     @Column
     private Long competition;
 
 
-    public UserSeason(int count, User player, int season, Long competition) {
+    public UserSeason(int count, User player, Long season, Long competition) {
         this. count = count;
         this.player = player;
         this.season = season;
